Move the ingest http handlers into the httpingestserver package
The handlers do not need to be in a separate package, and should be consistend with how the finder server is set up.
